Button Events will now work.
The buttonConfig.csv file can now be edited to any PROPER format and
will execute the function when all the buttons for that command are
pressed. The more buttons in a combo equals higher priority when
checking for the events. The implementationCheck() function will warn
against any functions that have been set to call but don't exist