Button Functionality Expanded
Button Functionality was expanded to allow the user to add a single
parameter to their event functions is 9_Event(param). This param will be
one of three values; "single", "double", or "long". Single occurs when
the button is released before the 300ms timer expires. Double occurs
when the button is pressed a second time before the 300ms timer expires
(does not have to be released in that time frame). Long occurs when the
button is held past for the full 300ms timer.